Fixed unit testing failures due to timechange.
This is temporary until i figure out a way to create unit tests which change based on whether the current time is in daylight savings or not.
The script functions as it should, but since the tests are testing against hard coded dates, it poses an issue only with testing.